LPS-28162 Implement undo functionality

+++ portal-impl/src/com/liferay/portlet/wiki/action/EditNodeAction.java
@@ -14,10 +14,13 @@
 
 package com.liferay.portlet.wiki.action;
 
+import com.liferay.portal.kernel.portlet.LiferayPortletConfig;
 import com.liferay.portal.kernel.servlet.SessionErrors;
+import com.liferay.portal.kernel.servlet.SessionMessages;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.ServiceContextFactory;
@@ -29,6 +32,9 @@ import com.liferay.portlet.wiki.model.WikiNode;
 import com.liferay.portlet.wiki.service.WikiNodeServiceUtil;
 import com.liferay.portlet.wiki.util.WikiCacheThreadLocal;
 import com.liferay.portlet.wiki.util.WikiCacheUtil;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -59,7 +65,15 @@ public class EditNodeAction extends PortletAction {
 				updateNode(actionRequest);
 			}
 			else if (cmd.equals(Constants.DELETE)) {
-				deleteNode(actionRequest);
+				deleteNode(
+					(LiferayPortletConfig)portletConfig, actionRequest, false);
+			}
+			else if (cmd.equals(Constants.MOVE_TO_TRASH)) {
+				deleteNode(
+					(LiferayPortletConfig)portletConfig, actionRequest, true);
+			}
+			else if (cmd.equals(Constants.RESTORE)) {
+				restoreNode(actionRequest);
 			}
 			else if (cmd.equals(Constants.SUBSCRIBE)) {
 				subscribeNode(actionRequest);
@@ -119,26 +133,60 @@ public class EditNodeAction extends PortletAction {
 			getForward(renderRequest, "portlet.wiki.edit_node"));
 	}
 
-	protected void deleteNode(ActionRequest actionRequest) throws Exception {
+	protected void deleteNode(
+			LiferayPortletConfig liferayPortletConfig,
+			ActionRequest actionRequest, boolean moveToTrash)
+		throws Exception {
+
 		long nodeId = ParamUtil.getLong(actionRequest, "nodeId");
 
 		String oldName = getNodeName(nodeId);
 
 		WikiCacheThreadLocal.setClearCache(false);
 
-		WikiNodeServiceUtil.deleteNode(nodeId);
+		if (moveToTrash) {
+			WikiNodeServiceUtil.moveEntryToTrash(nodeId);
+		}
+		else {
+			WikiNodeServiceUtil.deleteNode(nodeId);
+		}
 
 		WikiCacheUtil.clearCache(nodeId);
 
 		WikiCacheThreadLocal.setClearCache(true);
 
 		updatePreferences(actionRequest, oldName, StringPool.BLANK);
+
+		if (moveToTrash) {
+			Map<String, String[]> data = new HashMap<String, String[]>();
+
+			data.put("restoreEntryIds", new String[]{String.valueOf(nodeId)});
+
+			SessionMessages.add(
+				actionRequest,
+				liferayPortletConfig.getPortletId() +
+					SessionMessages.KEY_SUFFIX_DELETE_SUCCESS_DATA, data);
+
+			SessionMessages.add(
+				actionRequest,
+				liferayPortletConfig.getPortletId() +
+					SessionMessages.KEY_SUFFIX_HIDE_DEFAULT_SUCCESS_MESSAGE);
+		}
 	}
 
 	protected String getNodeName(long nodeId) throws Exception {
 		WikiNode node = WikiNodeServiceUtil.getNode(nodeId);
 
 		return node.getName();
+	}
+
+	protected void restoreNode(ActionRequest actionRequest) throws Exception {
+		long[] restoreEntryIds = StringUtil.split(
+			ParamUtil.getString(actionRequest, "restoreEntryIds"), 0L);
+
+		for (long restoreEntryId : restoreEntryIds) {
+			WikiNodeServiceUtil.restoreNodeFromTrash(restoreEntryId);
+		}
 	}
 
 	protected void subscribeNode(ActionRequest actionRequest) throws Exception {
